Move cluster API from test to lib
